sudo everything

+++ build-arch-base.sh
@@ -1,13 +1,18 @@
 #!/bin/sh
 
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root"
+   exit 1
+fi
+
 TIMESTAMP=$(date +%s)
 IMAGE_NAME="arch-base"
 
 rm -rf "${IMAGE_NAME}"
 mkdir "${IMAGE_NAME}"
 
-sudo pacstrap -c -d ${IMAGE_NAME} base || exit 1
+pacstrap -c -d ${IMAGE_NAME} base || exit 1
 
-sudo tar -C ${IMAGE_NAME} -c . | docker import - "${IMAGE_NAME}:${TIMESTAMP}"
+tar -C ${IMAGE_NAME} -c . | docker import - "${IMAGE_NAME}:${TIMESTAMP}"
 
 docker tag "${IMAGE_NAME}:${TIMESTAMP}" "${IMAGE_NAME}:latest"
